organize test folder
